Move config tests into test/untils/config
