+++ packages/astro/test/units/config/config-validate.test.js
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 import { z } from 'zod';
 import stripAnsi from 'strip-ansi';
-import { formatConfigErrorMessage } from '../dist/core/messages.js';
-import { validateConfig } from '../dist/core/config/index.js';
+import { formatConfigErrorMessage } from '../../../dist/core/messages.js';
+import { validateConfig } from '../../../dist/core/config/index.js';
 
 describe('Config Validation', () => {
 	it('empty user config is valid', async () => {
